chore: move action cell factory to shared components
